+++ src/main/kotlin/io/openfuture/chain/service/DefaultBlockService.kt
@@ -1,11 +1,10 @@
 package io.openfuture.chain.service
 
+import io.openfuture.chain.domain.transaction.RewardTransactionDto
 import io.openfuture.chain.entity.Block
 import io.openfuture.chain.entity.GenesisBlock
 import io.openfuture.chain.entity.MainBlock
-import io.openfuture.chain.entity.transaction.BaseTransaction
-import io.openfuture.chain.entity.transaction.TransferTransaction
-import io.openfuture.chain.entity.transaction.VoteTransaction
+import io.openfuture.chain.entity.transaction.*
 import io.openfuture.chain.exception.NotFoundException
 import io.openfuture.chain.repository.BlockRepository
 import io.openfuture.chain.repository.GenesisBlockRepository
@@ -19,7 +18,9 @@ class DefaultBlockService(
     private val mainBlockRepository: MainBlockRepository,
     private val genesisBlockRepository: GenesisBlockRepository,
     private val voteTransactionService: VoteTransactionService,
-    private val transferTransactionService: TransferTransactionService
+    private val transferTransactionService: TransferTransactionService,
+    private val delegateTransactionService: DelegateTransactionService,
+    private val rewardTransactionService: RewardTransactionService
 ) : BlockService {
 
     @Transactional(readOnly = true)
@@ -44,6 +45,8 @@ class DefaultBlockService(
 
     @Transactional
     override fun save(block: MainBlock): MainBlock {
+        rewardTransactionService.add(RewardTransactionDto(block.transactions.first() as RewardTransaction))
+
         val savedBlock = mainBlockRepository.save(block)
         val transactions = block.transactions
         for (transaction in transactions) {
@@ -61,6 +64,9 @@ class DefaultBlockService(
         when (tx) {
             is VoteTransaction -> voteTransactionService.toBlock(tx, block)
             is TransferTransaction -> transferTransactionService.toBlock(tx, block)
+            is DelegateTransaction -> delegateTransactionService.toBlock(tx, block)
+            is RewardTransaction -> rewardTransactionService.toBlock(tx, block)
+            else -> throw IllegalStateException("Unknown transaction type")
         }
     }