+++ src/main/resources/db/migration/V1_5__table_transactions.sql
@@ -1,7 +1,8 @@
 CREATE TABLE transactions (
   id                INTEGER PRIMARY KEY,
   timestamp         BIGINT         NOT NULL,
-  amount            BIGINT           NOT NULL,
+  amount            BIGINT         NOT NULL,
+  fee               BIGINT         NOT NULL,
   recipient_address VARCHAR        NOT NULL,
   sender_key        VARCHAR        NOT NULL,
   sender_address    VARCHAR        NOT NULL,
@@ -31,4 +32,8 @@ CREATE TABLE vote_transactions (
   id           INTEGER PRIMARY KEY REFERENCES transactions,
   vote_type_id INTEGER NOT NULL REFERENCES vote_types,
   delegate_key VARCHAR NOT NULL
+);
+--
+CREATE TABLE reward_transactions (
+  id INTEGER PRIMARY KEY REFERENCES transactions
 );