+++ src/test/kotlin/io/openfuture/chain/block/BlockCreationProcessorTests.kt
@@ -1,19 +1,17 @@
 package io.openfuture.chain.block
 
 import io.openfuture.chain.block.validation.BlockValidationProvider
+import io.openfuture.chain.component.converter.transaction.impl.RewardTransactionEntityConverter
 import io.openfuture.chain.component.node.NodeClock
 import io.openfuture.chain.config.ServiceTests
 import io.openfuture.chain.crypto.key.NodeKeyHolder
-import io.openfuture.chain.crypto.util.HashUtils
-import io.openfuture.chain.domain.block.BlockCreationEvent
 import io.openfuture.chain.domain.block.PendingBlock
 import io.openfuture.chain.domain.block.Signature
 import io.openfuture.chain.entity.Block
-import io.openfuture.chain.entity.Delegate
-import io.openfuture.chain.entity.GenesisBlock
 import io.openfuture.chain.entity.MainBlock
 import io.openfuture.chain.entity.transaction.BaseTransaction
 import io.openfuture.chain.entity.transaction.VoteTransaction
+import io.openfuture.chain.property.ConsensusProperties
 import io.openfuture.chain.service.BlockService
 import io.openfuture.chain.service.ConsensusService
 import io.openfuture.chain.service.DelegateService
@@ -22,7 +20,7 @@ import org.junit.Test
 import org.mockito.BDDMockito.given
 import org.mockito.Mock
 
-class BlockCreationProcessorTests: ServiceTests() {
+class BlockCreationProcessorTests : ServiceTests() {
 
     @Mock private lateinit var blockService: BlockService
     @Mock private lateinit var signatureCollector: SignatureCollector
@@ -31,15 +29,16 @@ class BlockCreationProcessorTests: ServiceTests() {
     @Mock private lateinit var consensusService: ConsensusService
     @Mock private lateinit var clock: NodeClock
     @Mock private lateinit var delegateService: DelegateService
+    @Mock private lateinit var rewardTransactionEntityConverter: RewardTransactionEntityConverter
+    @Mock private lateinit var consensusProperties: ConsensusProperties
 
     private lateinit var processor: BlockCreationProcessor
 
+
     @Before
     fun init() {
-        val block = createMainBlock()
-        given(blockService.getLastMain()).willReturn(block)
         processor = BlockCreationProcessor(blockService, signatureCollector, keyHolder, blockValidationService,
-            consensusService, clock, delegateService)
+            consensusService, clock, delegateService, rewardTransactionEntityConverter, consensusProperties)
     }
 
     @Test(expected = IllegalArgumentException::class)
@@ -71,23 +70,6 @@ class BlockCreationProcessorTests: ServiceTests() {
         processor.approveBlock(pendingBlock)
     }
 
-    @Test
-    fun fireBlockCreationShouldCreateMainBlock() {
-        val transactions = createTransactions()
-        val genesisBlock = createGenesisBlock()
-
-        given(keyHolder.getPublicKey()).willReturn(HashUtils.fromHexString("public_key"))
-        given(keyHolder.getPrivateKey()).willReturn(HashUtils.fromHexString("private_key"))
-
-        val delegate = Delegate(HashUtils.toHexString(keyHolder.getPublicKey()), "address", 1)
-        genesisBlock.activeDelegates = setOf(delegate)
-        val event = BlockCreationEvent(transactions)
-
-        given(blockService.getLastGenesis()).willReturn(genesisBlock)
-
-        processor.fireBlockCreation(event)
-    }
-
     private fun createPendingBlock(block: Block): PendingBlock {
         return PendingBlock(
             block,
@@ -104,23 +86,11 @@ class BlockCreationProcessorTests: ServiceTests() {
         createTransactions()
     )
 
-    private fun createGenesisBlock() = GenesisBlock(
-        ByteArray(1),
-        123,
-        "prev_block_hash",
-        1512345678L,
-        1,
-        setOf(
-            Delegate("public_key1", "host1", 1),
-            Delegate("public_key2", "host2", 2),
-            Delegate("public_key3", "host3", 3)
-        )
-    )
-
     private fun createTransactions(): MutableList<BaseTransaction> = mutableListOf(
         VoteTransaction(
             1500000000L,
             1000,
+            10,
             "recipient_address",
             "sender_key",
             "sender_address",
@@ -132,6 +102,7 @@ class BlockCreationProcessorTests: ServiceTests() {
         VoteTransaction(
             1500000001L,
             1002,
+            10,
             "recipient_address2",
             "sender_key2",
             "sender_address2",
@@ -141,4 +112,5 @@ class BlockCreationProcessorTests: ServiceTests() {
             "delegate_key2"
         )
     )
+
 }