+++ src/test/kotlin/io/openfuture/chain/block/validation/BlockValidationProviderTests.kt
@@ -37,6 +37,7 @@ class BlockValidationProviderTests : ServiceTests() {
                 VoteTransaction(
                     1500000000L,
                     1000,
+                    10,
                     "recipient_address",
                     "sender_key",
                     "sender_address",
@@ -48,6 +49,7 @@ class BlockValidationProviderTests : ServiceTests() {
                 VoteTransaction(
                     1500000001L,
                     1002,
+                    10,
                     "recipient_address2",
                     "sender_key2",
                     "sender_address2",
@@ -86,6 +88,7 @@ class BlockValidationProviderTests : ServiceTests() {
                 VoteTransaction(
                     1500000000L,
                     1000,
+                    10,
                     "recipient_address",
                     "sender_key",
                     "sender_address",
@@ -97,6 +100,7 @@ class BlockValidationProviderTests : ServiceTests() {
                 VoteTransaction(
                     1500000001L,
                     1002,
+                    10,
                     "recipient_address2",
                     "sender_key2",
                     "sender_address2",
@@ -127,6 +131,7 @@ class BlockValidationProviderTests : ServiceTests() {
                 VoteTransaction(
                     1500000000L,
                     1000,
+                    10,
                     "recipient_address",
                     "sender_key",
                     "sender_address",
@@ -138,6 +143,7 @@ class BlockValidationProviderTests : ServiceTests() {
                 VoteTransaction(
                     1500000001L,
                     1002,
+                    10,
                     "recipient_address2",
                     "sender_key2",
                     "sender_address2",