+++ CHANGELOG.md
@@ -5,6 +5,16 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Added
+- Fees to transactions
+- New type transaction: RewardTransaction. New entity, repository, service
+- Addition to block
+- Genesis address to properties
+- Constants fees and reward for transactions/block to properties
+- Verify the reward transaction
+
+### Changed
+- Structure of transactions
 
 ## [0.3.0] - 2018-07-13
 ### Added